requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==3.6.4.0
Brotli==1.1.0
celery==5.2.7
certifi==2019.3.9
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.2.1
crispy==0.7.4
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
daphne==3.0.2
defusedxml==0.7.1
Django==5.0.6
django-allauth==0.61.1
django-braces==1.15.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==4.3.0
django-embed-video==1.4.10
django-extensions==3.2.3
django-jalali==6.0.1
django-localflavor==3.1
django-parler==2.3
django-paypal==2.1
django-redisboard==8.4.0
django-rosetta==0.10.0
djangorestframework==3.15.2
easy-thumbnails==2.8.5
fabio==2024.4.0
flower==2.0.1
fonttools==4.51.0
h5py==3.11.0
hdf5plugin==4.4.0
html5lib==1.1
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
jdatetime==4.1.1
kiwisolver==1.4.5
kombu==5.3.5
lxml==5.2.2
matplotlib==3.9.0
msgpack==1.0.8
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pep8==1.7.1
pillow==10.2.0
polib==1.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.5.0
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pymemcache==4.0.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyphen==0.14.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
python-decouple==3.1
python-stdnum==1.20
python3-openid==3.2.0
pytube==15.0.0
pytz==2018.5
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
rosetta==0.3
scipy==1.13.0
service-identity==24.1.0
silx==2.1.0
six==1.16.0
sqlparse==0.5.0
stripe==4.0.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.10.0
ua-parser==0.18.0
urllib3==1.24.2
user-agent==0.1.10
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
youtube-dl==2021.12.17
zope.interface==7.0.3
zopfli==0.2.3
